+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 
 - Attacks
   - EAPOL Logoff Attack
-
+  - SAE Flood Attack
+  
 ### Changed
 
 - Enable Cardputer's LED in config - @tototo31